pyproject.toml
[project]
name = "medium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "celery>=5.5.3",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "python-decouple>=3.8",
    "python-jose>=3.5.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
